Conflicts:
	handlers/PostHandler.js

+++ handlers/AccountHandler.js
@@ -2,6 +2,7 @@
 var Account = require('../models/Account')
 var config = require('../config.js')
 var r = require('rethinkdb')
+var auth = require('../auth.js')
 
 var AccountHandler = function () {
   this.createAccount = handleCreateAccountRequest
@@ -19,8 +20,11 @@ r.connect( {host: config.rethinkdb.host, port: config.rethinkdb.port}, function(
 // called when a user logs in, add userId to DB if not present
 // create Account object, add data to DB using thinky
 function handleCreateAccountRequest (req, res) {
+  console.dir(req.body)
+  if (!auth.assertHasUser(req)) return
+
   // create Account object
-  var account = new Account({userId: req.headers.userid})
+  var account = new Account({userId: req.headers.userid, name: req.body.name})
 
   // use Thinky to save Account data
   account.save().then(function (result) {
@@ -37,7 +41,7 @@ function handleGetAccountRequest (req, res) {
   for (var q in req.query) {
     if (req.query.hasOwnProperty(q)) {
       queried = true
-      r.db(config.rethinkdb.db).table('users').filter(r.row(q).eq(req.query[q])).run(
+      r.db(config.rethinkdb.db).table('users').filter(r.row(q).eq(parseInt(req.query[q]))).run(
           connection, function (err, cursor) {
             if (err) throw err
             cursor.toArray(function (err, result) {
@@ -64,21 +68,23 @@ function handleUpdateAccountRequest (req, res) {
   console.log('handleUpdateAccountRequest called with ' + JSON.stringify(req.route))
 }
 function handleDeleteAccountRequest (req, res) {
-  console.log('handleDeleteAccountRequest called with ' + JSON.stringify(req.route))
+  if (!auth.assertHasUser(req)) return
 
   r.db(config.rethinkdb.db).table('users').filter(req.headers.userid).delete().run(
-         connection, function(err, cursor){
-          if (err) throw err
-        }).then(function(result) {
-           res.json({
-               result: result
-           })
-       })
+    connection, function(err, cursor) {
+      if (err) throw err
+    }
+  ).then(function(result) {
+    res.json({
+      result: result
+    })
+  })
 
   r.db(config.rethinkdb.db).table('favorites').filter({"userId": req.headers.userid}).delete().run(
-         connection, function(err, cursor){
-          if (err) throw err
-        })
-  }
+    connection, function(err, cursor) {
+      if (err) throw err
+    }
+  )
+}
 
 module.exports = AccountHandler
